viz/requirements.txt
meerkat-ml
transformers
pyarrow
tqdm
rich
faiss-cpu